Amend initialize / finalize through v8
Also Expose .init()-factory to get a new reference of the cpp Ncpy class.
Adapt examples.

This primarily changes the JS logic at the interface of calling the cpp object but not exposing it as singleton. This makes it necessary, that cpp land handles memory lifetime of the python interpreter. Also be aware that this behaviour supports just one interpreter.

TODO: in the future we might support several by calls to to